move index.js
