+++ pages/_app.tsx
@@ -1,7 +1,46 @@
 import type { AppProps } from 'next/app';
+import React from 'react';
+import ReactDOM from 'react-dom';
+
+import { CssBaseline } from '@material-ui/core';
+import { MuiThemeProvider } from '@material-ui/core/styles';
+
+import AppStateProvider, { useAppState } from '../src/state';
+import ErrorDialog from '../src/components/ErrorDialog/ErrorDialog';
+import theme from '../src/theme';
+import '../src/types';
+import { ChatProvider } from '../src/components/ChatProvider';
+import { VideoProvider } from '../src/components/VideoProvider';
+import useConnectionOptions from '../src/utils/useConnectionOptions/useConnectionOptions';
+import UnsupportedBrowserWarning from '../src/components/UnsupportedBrowserWarning/UnsupportedBrowserWarning'
+
+const VideoApp = (props) => {
+    const { error, setError } = useAppState();
+    const connectionOptions = useConnectionOptions();
+
+    return (
+        <VideoProvider options={connectionOptions} onError={setError}>
+            <ErrorDialog dismissError={() => setError(null)} error={error} />
+            <ChatProvider>
+                {props.children}
+            </ChatProvider>
+        </VideoProvider>
+    );
+};
 
 function App({ Component, pageProps }: AppProps) {
     return (
-        <Component {...pageProps} />
+        <MuiThemeProvider theme={theme}>
+            <CssBaseline />
+            <UnsupportedBrowserWarning>
+                <AppStateProvider>
+                    <VideoApp>
+                        <Component {...pageProps} />
+                    </VideoApp>
+
+                </AppStateProvider>
+
+            </UnsupportedBrowserWarning>
+        </MuiThemeProvider>
     );
 }
